move callback group parsing to ObeliskNode.py, add generic lifecycle callback implementations
